requirements.txt
django==1.8.3
django-bootstrap3==6.2.0
django-crispy-forms==1.5.1
djangorestframework==3.2.2
djangorestframework-filters==0.4.0